@@ -1,7 +1,7 @@
 source 'http://rubygems.org'
-ruby '2.3.3'
+ruby '2.7.2'
 gem 'sinatra'
-gem 'activerecord', '~> 4.2', '>= 4.2.6', :require => 'active_record'
+gem 'activerecord', '~> 5.2', '>= 5.2.0', :require => 'active_record'
 gem 'sinatra-activerecord', :require => 'sinatra/activerecord'
 gem 'rake'
 gem 'sinatra-flash'
@@ -11,13 +11,12 @@ gem 'shotgun'
 gem 'pony'
 gem 'bcrypt'
 gem "tux"
+gem "mysql2"
 
-group :production do
-  gem 'pg', '0.15'
-end
 
-group :development do 
-	gem 'sqlite3'
+
+
+group :production, :development do
 	gem 'pry'
 end
 
@@ -27,4 +26,5 @@ group :test do
   gem 'capybara'
   gem 'rack-test'
   gem 'database_cleaner', git: 'https://github.com/bmabey/database_cleaner.git'
+  gem 'sqlite3'
 end
