@@ -1,9 +1,12 @@
-class CreateUsers < ActiveRecord::Migration
+class CreateUsers < ActiveRecord::Migration[5.2]
   def change
     create_table :users do |t|
-      t.string :name
+      t.string :fname
+      t.string :lname
       t.string :email
+      t.string :username
       t.string :password_digest
+
       t.timestamps null: false
     end
   end
