@@ -25,7 +25,7 @@ class UsersController < ApplicationController
   # GET: /let the user to go for the sign-in page --done
   get "/signin" do
     if signed_in?
-      redirect '/todos'
+      redirect '/listings'
     else
       erb :"/users/signin.html"
     end
@@ -34,7 +34,7 @@ class UsersController < ApplicationController
   # GET: /let the user go for the sign-up page --done
   get "/signup" do
     if signed_in?
-      redirect '/todos'
+      redirect '/listings'
     else
       erb :"/users/new.html"
     end
@@ -42,10 +42,10 @@ class UsersController < ApplicationController
 
   # POST: /send the sign-in info to the server and let the user to login
   post "/signin" do
-    @user = User.find_by(:name => params[:name])
+    @user = User.find_by(:username => params[:username])
     if @user && @user.authenticate(params[:password])
       session[:user_id] = @user.id
-      redirect '/todos'
+      redirect '/listings'
     else
       redirect "/signup"
     end
@@ -53,17 +53,19 @@ class UsersController < ApplicationController
   #POST:/send the signup infor to the serverand let the user to create account
   post "/signup" do
     # if one of the entry field is empty direct to the signup page
-    if params[:name].empty? || params[:email].empty? || params[:password].empty?
+    if params[:fname].empty? || params[:lname].empty? || params[:username].empty? || params[:email].empty? || params[:password].empty?
       redirect "/signup"
     else
       #else create a new instance of user using params
       # set session[:user_id] to newly created user id
-      #finally redirect the user to the todos list page
+      #finally redirect the user to the listings list page
       # binding.pry
-      @user = User.create(:name => params[:name], :email => params[:email], :password => params[:password])
-      # @user.save
+      @user = User.create(:fname => params[:fname], :lname => params[:lname], :username => params[:username], :email => params[:email], :password => params[:password])
+      # @user = User.create(params)
+
+      @user.save
       session[:user_id] = @user.id
-      redirect "/todos"
+      redirect "/listings"
     end
   end
   get "/signout" do
@@ -92,12 +94,12 @@ class UsersController < ApplicationController
   patch '/users/:id' do
     # binding.pry
     if signed_in?
-      if params[:name].empty?
+      if params[:fname].empty?
         redirect "/users/#{params[:id]}/edit"
       else
         @user = User.find_by_id(params[:id])
         if @user == current_user
-          if @user.update(:name => params[:name], :email => params[:email])
+          if @user.update(:fname => params[:fname], :lname => params[:lname], :username => params[:username], :email => params[:email])
             redirect to "/users/#{@user.id}"
           else
           redirect to "/users/#{@user.id}/edit"
