@@ -1,5 +1,6 @@
 ENV['SINATRA_ENV'] ||= "development"
 require 'bundler/setup'
+require 'mysql2'
 
 Bundler.require(:default, ENV['SINATRA_ENV'])
 
@@ -7,13 +8,16 @@ ActiveRecord::Base.logger = Logger.new(STDOUT)
 
 configure :development do
   ActiveRecord::Base.establish_connection(
-    :adapter => "sqlite3",
-    :database => "db/#{ENV['SINATRA_ENV']}.sqlite"
+    :adapter => "mysql2",
+    :encoding => "utf8",
+    :database => "dev_todo_app",
+    :username => "root",
+    :password => "Palo5non!"
   )
 end
 
 configure :production do
-  ActiveRecord::Base.establish_connection(ENV['DATABASE_URL'] || 'postgres://localhost/mydb')
+  ActiveRecord::Base.establish_connection(ENV['DATABASE_URL'] || 'mysql://localhost/mydb')
 end
 
 require_all 'app'
